refactor: dto 패키지 구조 리팩토링
- 기존엔 도메인별로 패키지를 만들어서 구분해 줬는데, request/response 패키지만 만들어서 정리
