+++ src/main/java/com/peeerr/climbing/controller/CategoryController.java
@@ -1,9 +1,9 @@
 package com.peeerr.climbing.controller;
 
-import com.peeerr.climbing.dto.ApiResponse;
-import com.peeerr.climbing.dto.category.CategoryCreateRequest;
-import com.peeerr.climbing.dto.category.CategoryEditRequest;
-import com.peeerr.climbing.dto.category.CategoryResponse;
+import com.peeerr.climbing.dto.common.ApiResponse;
+import com.peeerr.climbing.dto.request.CategoryCreateRequest;
+import com.peeerr.climbing.dto.request.CategoryEditRequest;
+import com.peeerr.climbing.dto.response.CategoryResponse;
 import com.peeerr.climbing.service.CategoryService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
