+++ src/main/java/com/peeerr/climbing/dto/common/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.peeerr.climbing.dto;
+package com.peeerr.climbing.dto.common;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
