+++ src/main/java/com/peeerr/climbing/dto/request/CategoryCreateRequest.java
@@ -1,4 +1,4 @@
-package com.peeerr.climbing.dto.category;
+package com.peeerr.climbing.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -10,14 +10,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class CategoryEditRequest {
+public class CategoryCreateRequest {
 
     @Size(min = 1, max = 20)
     @NotBlank(message = "카테고리명을 입력해 주세요.")
     private String categoryName;
 
-    public static CategoryEditRequest of(String categoryName) {
-        return new CategoryEditRequest(categoryName);
+    public static CategoryCreateRequest of(String categoryName) {
+        return new CategoryCreateRequest(categoryName);
     }
 
 }
