+++ src/main/java/com/peeerr/climbing/dto/request/PostCreateRequest.java
@@ -1,4 +1,4 @@
-package com.peeerr.climbing.dto.post;
+package com.peeerr.climbing.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
