+++ src/main/java/com/peeerr/climbing/repository/CustomPostRepository.java
@@ -1,8 +1,8 @@
 package com.peeerr.climbing.repository;
 
 import com.peeerr.climbing.entity.Post;
-import com.peeerr.climbing.dto.post.PostSearchCondition;
-import com.peeerr.climbing.dto.post.PostResponse;
+import com.peeerr.climbing.dto.request.PostSearchCondition;
+import com.peeerr.climbing.dto.response.PostResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
