+++ src/test/java/com/peeerr/climbing/integration/docs/CommentDocTest.java
@@ -13,8 +13,8 @@ import com.peeerr.climbing.entity.Post;
 import com.peeerr.climbing.repository.PostRepository;
 import com.peeerr.climbing.entity.Member;
 import com.peeerr.climbing.repository.MemberRepository;
-import com.peeerr.climbing.dto.comment.CommentCreateRequest;
-import com.peeerr.climbing.dto.comment.CommentEditRequest;
+import com.peeerr.climbing.dto.request.CommentCreateRequest;
+import com.peeerr.climbing.dto.request.CommentEditRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
