+++ src/test/java/com/peeerr/climbing/integration/docs/MemberDocTest.java
@@ -24,9 +24,9 @@ import com.peeerr.climbing.repository.LikeRepository;
 import com.peeerr.climbing.repository.PostRepository;
 import com.peeerr.climbing.entity.Member;
 import com.peeerr.climbing.repository.MemberRepository;
-import com.peeerr.climbing.dto.member.MemberCreateRequest;
-import com.peeerr.climbing.dto.member.MemberEditRequest;
-import com.peeerr.climbing.dto.member.MemberLoginRequest;
+import com.peeerr.climbing.dto.request.MemberCreateRequest;
+import com.peeerr.climbing.dto.request.MemberEditRequest;
+import com.peeerr.climbing.dto.request.MemberLoginRequest;
 import com.peeerr.climbing.security.MemberPrincipal;
 import com.peeerr.climbing.service.MemberService;
 import org.junit.jupiter.api.BeforeEach;
