+++ src/test/java/com/peeerr/climbing/repository/CustomPostRepositoryImplTest.java
@@ -2,8 +2,8 @@ package com.peeerr.climbing.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.peeerr.climbing.dto.post.PostResponse;
-import com.peeerr.climbing.dto.post.PostSearchCondition;
+import com.peeerr.climbing.dto.response.PostResponse;
+import com.peeerr.climbing.dto.request.PostSearchCondition;
 import com.peeerr.climbing.entity.Post;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import java.util.Optional;
