+++ src/test/java/com/peeerr/climbing/service/CategoryServiceTest.java
@@ -2,9 +2,9 @@ package com.peeerr.climbing.service;
 
 import com.peeerr.climbing.entity.Category;
 import com.peeerr.climbing.repository.CategoryRepository;
-import com.peeerr.climbing.dto.category.CategoryCreateRequest;
-import com.peeerr.climbing.dto.category.CategoryEditRequest;
-import com.peeerr.climbing.dto.category.CategoryResponse;
+import com.peeerr.climbing.dto.request.CategoryCreateRequest;
+import com.peeerr.climbing.dto.request.CategoryEditRequest;
+import com.peeerr.climbing.dto.response.CategoryResponse;
 import com.peeerr.climbing.exception.already.AlreadyExistsCategoryException;
 import com.peeerr.climbing.exception.notfound.CategoryNotFoundException;
 import org.junit.jupiter.api.DisplayName;
