+++ src/test/java/com/peeerr/climbing/service/MemberServiceTest.java
@@ -7,8 +7,8 @@ import static org.mockito.BDDMockito.then;
 
 import com.peeerr.climbing.entity.Member;
 import com.peeerr.climbing.repository.MemberRepository;
-import com.peeerr.climbing.dto.member.MemberCreateRequest;
-import com.peeerr.climbing.dto.member.MemberEditRequest;
+import com.peeerr.climbing.dto.request.MemberCreateRequest;
+import com.peeerr.climbing.dto.request.MemberEditRequest;
 import com.peeerr.climbing.exception.already.AlreadyExistsCategoryException;
 import com.peeerr.climbing.exception.notfound.CategoryNotFoundException;
 import com.peeerr.climbing.exception.AccessDeniedException;
