+++ src/test/java/com/peeerr/climbing/service/PostServiceTest.java
@@ -5,11 +5,11 @@ import com.peeerr.climbing.repository.CategoryRepository;
 import com.peeerr.climbing.entity.Post;
 import com.peeerr.climbing.repository.PostRepository;
 import com.peeerr.climbing.entity.Member;
-import com.peeerr.climbing.dto.post.PostCreateRequest;
-import com.peeerr.climbing.dto.post.PostEditRequest;
-import com.peeerr.climbing.dto.post.PostSearchCondition;
-import com.peeerr.climbing.dto.post.PostResponse;
-import com.peeerr.climbing.dto.post.PostDetailResponse;
+import com.peeerr.climbing.dto.request.PostCreateRequest;
+import com.peeerr.climbing.dto.request.PostEditRequest;
+import com.peeerr.climbing.dto.request.PostSearchCondition;
+import com.peeerr.climbing.dto.response.PostResponse;
+import com.peeerr.climbing.dto.response.PostDetailResponse;
 import com.peeerr.climbing.exception.notfound.CategoryNotFoundException;
 import com.peeerr.climbing.exception.AccessDeniedException;
 import org.junit.jupiter.api.DisplayName;
